Add the concept of category to Event's
Introduce a new type Category (uint8) and add a`cat` field to `Event`
structure.
If `Category` is not called, Event.cat is thus the zero-value of
`Category`, corresponding to NC (NotCategorized), specically handled in
`Event.Msg` so as to make the concept of Category 100% opt-in for client
code.

However if `Event.Category` is called, Event.cat is set so as to check,
in `Event.Msg`, if log level assigned to this category allows the event
to be logged. If that is the case, the event will be logged with an
extra key `category`, with as value, the category name, if prealably
set in client code.

Add, in `category.go` some functions to manage the category-based log
levels and category names.